:bug: [#954] fixed bug that would display toelichting block when canceling confirm
